+++ src/main/java/ua/algorithms/indicator/InsertIndicator.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class InsertIndicator implements Indicator {
     @Override
-    public int calculate(int id, IndexBlock block) {
+    public int calculate(long id, IndexBlock block) {
         List<IndexRecord> records = block.getRecords();
         IndexRecord first = records.get(0);
         if (records.size() == 1) {
